requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
websockets==12.0
iqoptionapi==0.5
pandas==2.2.1
numpy==1.26.4
ta==0.11.0
Pillow==10.2.0
pydantic==2.6.3
pydantic-settings==2.2.1
